rename : 기존 menu 명칭 document로 수정
