+++ style.css
@@ -112,8 +112,13 @@ li {
 }
 
 .textEditor {
+
     width: 500px;
     height: 500px;
+}
+
+.textEditor:focus {
+    outline: none;
 }
 
 .textEditor p,
